requirements.txt
boto3==1.24.33
botocore==1.27.33
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
idna==3.3
jmespath==1.0.1
msal==1.18.0
Office365-REST-Python-Client==2.3.13
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.10
